[test] Also look for compressed versions of a requested archive

+++ core/tests/helpers.py
@@ -3,6 +3,8 @@ import os
 import shutil
 import tempfile
 
+from ratarmountcore.compressions import COMPRESSION_FORMATS, find_available_backend, libarchive  # noqa: E402
+
 
 def find_test_file(relativePathOrName):
     for i in range(3):
@@ -16,5 +18,23 @@ def find_test_file(relativePathOrName):
 def copy_test_file(relativePathOrName):
     with tempfile.TemporaryDirectory() as folder:
         path = os.path.join(folder, os.path.basename(relativePathOrName))
-        shutil.copy(find_test_file(relativePathOrName), path)
+        if os.path.isfile(find_test_file(relativePathOrName)):
+            shutil.copy(find_test_file(relativePathOrName), path)
+        else:
+            for format_id, format_info in COMPRESSION_FORMATS.items():
+                backend = find_available_backend(format_id)
+                if not backend:
+                    continue
+
+                for extension in format_info.extensions:
+                    found_path = find_test_file(relativePathOrName + '.' + extension)
+                    if not os.path.isfile(found_path):
+                        continue
+
+                    with open(found_path, 'rb') as file, contextlib.suppress(Exception), backend.open(
+                        file
+                    ) as decompressed, open(path, 'wb') as target:
+                        shutil.copyfileobj(decompressed, target)
+                        break
+
         yield path
